eth/catalyst: move stateless api, nicer checks
